KimJaewoo/requirements.txt
# Core Dependencies
torch>=2.0.0

# LangChain & Vector DB
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
langchain-openai>=0.0.5
chromadb>=0.4.22

# Embedding Model
sentence-transformers>=2.2.2
huggingface-hub>=0.19.0

# OpenAI (for GPT-4o)
openai>=1.0.0

# Utils
python-dotenv>=1.0.0